Removes role=“alert” and aria-atomic="true” from alert box on forms.
For information-only boxes like on this page, remove role=“alert” and
aria-atomic="true”. They are intended for time sensitive error messages, things
like that. They tell the screen reader to read the content aloud as important,
and again whenever it changes, which is rather obnoxious for users if its not
necessary.